Refactor: 라운지 페이지 동적 라우팅 처리
